Create a minimal reproducible example & Add environs
Environment variables are taken from the file .env.
For this i use the library environs.

Bot.py is responsible for starting the bot. This is the main file.

Also add user_registration.py: it will register the user.
For example, who is the user: a student or a teacher, what group does he have etc.