Finish user_registration
How registration works:
	Who are u? --->
	Enter group name or name of the teacher --->
	If name in the list, then show the menu

FSM will allow u to remember the users choice and allow u to get rid of errors.
I dont split FSM into multiple FSM. It is one for the whole bot.

Register handlers from user_registration.py, create file full_names.py.
It contains a list of all groups and teachers.
So that the check "is it in the list" is faster, i make the list a dict.

Created keyboards: for student or teacher selection, for main menu