Add a function to get schedule
Schedule is stored in Google Drive, in specific/root folder
This folder contains months. For example, 09 September. These months are also folders.
Links to Google Spreadsheets are stored in the month. That is days from 1 to 29 or 30 etc.
Structure example:

root/
├─ 09 September/
│  ├─ 01 link
│  ├─ 02 link
│  ├─ 03 link
│  ├─ ...
├─ 10 November/
│  ├─ 01 link
│  ├─ 02 link
│  ├─ 03 link
│  ├─ ...

Work with Google Drive using pydrive.
Open links and get data using gspread.
Google API Auth via service account.

Added a cache in the form a dictionary, because there is not much data,
and they are not important. For each group and each teacher, I add a weekly
schedule to the cache.